### 2022-09-23 #4779 @kyleniemeyer
Hi @kathlandgren, we will use this issue to assign an editor and find reviewers.

Unfortunately, we do not have any editors in this area available to handle this right now, so I need to put this on our wait list until someone can take it on. In the meantime, any reviewer suggestions would be welcome.

Also, it looks like some of your references in the paper are missing journal names, likely due to macros that our paper-building system doesn't recognize. Can you replace those with typed-out journal names?
