+++ paper/paper.bib
@@ -12,7 +12,7 @@
 @article{Bell:2018,
        author = {{Bell}, Taylor J. and {Cowan}, Nicolas B.},
         title = "{Increased Heat Transport in Ultra-hot Jupiter Atmospheres through H$_{2}$ Dissociation and Recombination}",
-      journal = {\apjl},
+      journal = {The Astrophysical Journal Letters},
      keywords = {methods: analytical, methods: numerical, planets and satellites: atmospheres, planets and satellites: gaseous planets, Astrophysics - Earth and Planetary Astrophysics},
          year = 2018,
         month = apr,
@@ -58,7 +58,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Menou:2009,
        author = {{Menou}, Kristen and {Rauscher}, Emily},
         title = "{Atmospheric Circulation of Hot Jupiters: A Shallow Three-Dimensional Model}",
-      journal = {\apj},
+      journal = {The Astrophysical Journal},
      keywords = {infrared: general, infrared: stars, planetary systems, stars: atmospheres, turbulence, Astrophysics},
          year = 2009,
         month = jul,
@@ -78,7 +78,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Kataria:2016,
        author = {{Kataria}, Tiffany and {Sing}, David K. and {Lewis}, Nikole K. and {Visscher}, Channon and {Showman}, Adam P. and {Fortney}, Jonathan J. and {Marley}, Mark S.},
         title = "{The Atmospheric Circulation of a Nine-hot-Jupiter Sample: Probing Circulation and Chemistry over a Wide Phase Space}",
-      journal = {\apj},
+      journal = {The Astrophysical Journal},
      keywords = {atmospheric effects, methods: numerical, planets and satellites: general, Astrophysics - Earth and Planetary Astrophysics},
          year = 2016,
         month = apr,
@@ -97,7 +97,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Parmentier:2013,
        author = {{Parmentier}, Vivien and {Showman}, Adam P. and {Lian}, Yuan},
         title = "{3D mixing in hot Jupiters atmospheres. I. Application to the day/night cold trap in HD 209458b}",
-      journal = {\aap},
+      journal = {Astronomy and Astrophysics},
      keywords = {planets and satellites: atmospheres, methods: numerical, diffusion, Astrophysics - Earth and Planetary Astrophysics},
          year = 2013,
         month = oct,
@@ -115,7 +115,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Cooper:2006,
        author = {{Cooper}, Curtis S. and {Showman}, Adam P.},
         title = "{Dynamics and Disequilibrium Carbon Chemistry in Hot Jupiter Atmospheres, with Application to HD 209458b}",
-      journal = {\apj},
+      journal = {The Astrophysical Journal},
      keywords = {Atmospheric Effects, Methods: Numerical, Stars: Planetary Systems, Planets and Satellites: General, planets and satellites: individual (HD 09458b), Astrophysics},
          year = 2006,
         month = oct,
@@ -133,7 +133,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Dobbs-Dixon:2013,
        author = {{Dobbs-Dixon}, Ian and {Agol}, Eric},
         title = "{Three-dimensional radiative-hydrodynamical simulations of the highly irradiated short-period exoplanet HD 189733b}",
-      journal = {\mnras},
+      journal = {Monthly Notices of the Royal Astronomical Society},
[…]
