### 2022-09-30 #4779 @kathlandgren
Thank you, @kyleniemeyer!

Here are some reviewer suggestions:

- harpolea
- mikaem
- ali-ramadhan
- zingale
- JoshKarpel
- ysimillides

I have replaced the macros with typed-out journal names. Some references do not have associated journals (a dissertation and a technical note).
