### 2022-10-21 #4872 @editorialbot
[REVIEW]: SWAMPE: A Shallow-Water Atmospheric Model in Python for Exoplanets
**Submitting author:** @kathlandgren (<a href="http://orcid.org/0000-0001-6029-5216">Ekaterina Landgren</a>)
**Repository:** https://github.com/kathlandgren/SWAMPE
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
