+++ paper/paper.bib
@@ -293,3 +293,22 @@ archivePrefix = {arXiv},
   pages={2346--2360},
   year={2001}
 }
+
+@ARTICLE{Burns:2020,
+       author = {{Burns}, Keaton J. and {Vasil}, Geoffrey M. and {Oishi}, Jeffrey S. and {Lecoanet}, Daniel and {Brown}, Benjamin P.},
+        title = "{Dedalus: A flexible framework for numerical simulations with spectral methods}",
+      journal = {Physical Review Research},
+     keywords = {Astrophysics - Instrumentation and Methods for Astrophysics, Physics - Computational Physics, Physics - Fluid Dynamics},
+         year = 2020,
+        month = apr,
+       volume = {2},
+       number = {2},
+          eid = {023068},
+        pages = {023068},
+          doi = {10.1103/PhysRevResearch.2.023068},
+archivePrefix = {arXiv},
+       eprint = {1905.10388},
+ primaryClass = {astro-ph.IM},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2020PhRvR...2b3068B},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
