### 2022-11-01 #4872 @imalsky
Thanks! I opened a review. My apologies if I get any of the process wrong,
this is my first JOSS review.

On Mon, Oct 31, 2022 at 7:40 AM Dan Foreman-Mackey ***@***.***>
wrote:

> @imalsky <https://github.com/imalsky> — A reminder to keep this on your
> radar. Please generate your checklist and start taking a look as soon as
> you can. Thanks!!
>
