+++ docs/source/notebooks/hot_Jupiter_reference_example/readme.md
@@ -1,6 +1,10 @@
-This folder contains the scripts used to generate three simulations of a hot Jupiter. The scripts include both planetary and simulation parameters.
+## Reference data
+
+This folder contains the scripts used to generate three simulations of a hot Jupiter which appear in the [QuickStart notebook](https://github.com/kathlandgren/SWAMPE/blob/validation/docs/source/notebooks/QuickStart.ipynb). The scripts include both planetary and simulation parameters. 
 
 Each simulation was run for 1500 h. The reference data provided here are snapshots at 1500 h. The simulations for taurad=0.1 days equilibrate at around 200 hours. The simulations for taurad=1 day and taurad=10 days equilibrate after approximately 1000 hours.
 
 These simulations took between 21 and 24 hours to complete while running simultaneously on a single core.
 
+The resulting geopotential profiles are illustrated below:
+![geopotential profiles for a hot Jupiter planet for three radiative timescales.](https://github.com/kathlandgren/SWAMPE/blob/validation/docs/_static/timescale_example-2.png)
