+++ paper/paper.bib
@@ -312,3 +312,14 @@ archivePrefix = {arXiv},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2020PhRvR...2b3068B},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+@article{Dunne:2020,
+  title={The GFDL Earth System Model version 4.1 (GFDL-ESM 4.1): Overall coupled model description and simulation characteristics},
+  author={Dunne, JP and Horowitz, LW and Adcroft, AJ and Ginoux, P and Held, IM and John, JG and Krasting, JP and Malyshev, S and Naik, V and Paulot, F and others},
+  journal={Journal of Advances in Modeling Earth Systems},
+  volume={12},
+  number={11},
+  pages={e2019MS002015},
+  year={2020},
+  publisher={Wiley Online Library}
+}
