+++ README.md
@@ -33,7 +33,11 @@ Please see the LICENSE file for more information.
 Support
 -------
 
-Please reach out with any feedback, bug reports, or questions to ek672@cornell.edu.
+
+To report a bug or request a new feature, please open a [new issue](https://github.com/kathlandgren/SWAMPE/issues/new).
+
+If you would like to get more actively involved in SWAMPE development, you are invited to get in touch ek672@cornell.edu,
+and I will be happy to answer questions about the codebase.
 
 [![License](https://img.shields.io/badge/License-BSD_3--Clause-blue.svg)](https://opensource.org/licenses/BSD-3-Clause)
 [![Github Actions CI](https://github.com/kathlandgren/SWAMPE/actions/workflows/main.yml/badge.svg)](https://github.com/kathlandgren/SWAMPE/actions/workflows/main.yml)
