+++ docs/source/index.rst
@@ -8,7 +8,7 @@ SWAMPE's documentation
 
 SWAMPE is a 2D shallow-water general circulation model designed for simulating
 exoplanet atmospheres. SWAMPE is a fully in Python implementation of a spectral
-algorithm described in `Hack and Jakob (1992) <"https://opensky.ucar.edu/islandora/object/technotes:112">`_
+algorithm described in `Hack and Jakob (1992) <https://opensky.ucar.edu/islandora/object/technotes:112>`_
 and previously available only in Fortran.
 
 SWAMPE's features currently include:
@@ -29,7 +29,7 @@ SWAMPE is available under the BSD 3-Clause License.
 Contributing
 ==============
 
-To report a bug or request a new feature, please open a `new issue <"https://github.com/kathlandgren/SWAMPE/issues/new">`_.
+To report a bug or request a new feature, please open a `new issue <https://github.com/kathlandgren/SWAMPE/issues/new>`_.
 
 If you would like to get more actively involved in SWAMPE development, you are invited to get in touch ek672-at-cornell-dot-edu,
 and I will be happy to help you understand the codebase as you work on your contribution.
