### 2022-11-30 #4872 @mark-hammond
Thank you for addressing the comments, I have ticked all the boxes now and closed the issue thread on the other repo. This is a great piece of code, thank you for sharing it in this way!
