### 2022-12-01 #4872 @dfm
Awesome!! Thanks @mark-hammond and @imalsky for your constructive reviews! I really appreciate all the time that you volunteered to this process.

@kathlandgren — I have a couple of final edits to do then I'll have some quick tasks for you before we publish. I'm totally swamped today, but I'll get to this before the end of the week. Thanks for your patience!
