### 2022-12-02 #4872 @imalsky
Thanks for inviting me! It was my pleasure

On Thu, Dec 1, 2022 at 3:20 AM Dan Foreman-Mackey ***@***.***>
wrote:

> Awesome!! Thanks @mark-hammond <https://github.com/mark-hammond> and
> @imalsky <https://github.com/imalsky> for your constructive reviews! I
> really appreciate all the time that you volunteered to this process.
>
> @kathlandgren <https://github.com/kathlandgren> — I have a couple of
> final edits to do then I'll have some quick tasks for you before we
> publish. I'm totally swamped today, but I'll get to this before the end of
> the week. Thanks for your patience!
>
