### 2022-12-05 350d646
edits for JOSS submission (#7)
