+++ paper/paper.bib
@@ -6,7 +6,8 @@
   number={C3},
   pages={5733--5752},
   year={1997},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1029/96jc02776}
 }
 
 @article{Bell:2018,
@@ -37,7 +38,7 @@ archivePrefix = {arXiv},
 }
 
 @book{Langton:2008,
-  title={Atmospheric dynamics on strongly irradiated Jovian planets},
+  title={Atmospheric dynamics on strongly irradiated {Jovian} planets},
   author={Langton, Jonathan},
   year={2008},
   publisher={University of California, Santa Cruz}
@@ -52,7 +53,8 @@ archivePrefix = {arXiv},
   number={1},
   pages={52},
   year={2016},
-  publisher={IOP Publishing}
+  publisher={IOP Publishing},
+  doi={10.3847/0004-637x/829/1/52}
 }
 
 @ARTICLE{Menou:2009,
@@ -156,7 +158,8 @@ archivePrefix = {arXiv},
   number={1},
   pages={12},
   year={2017},
-  publisher={IOP Publishing}
+  publisher={IOP Publishing},
+  doi={10.3847/1538-4365/aa7a06}
 }
 
 
@@ -168,7 +171,8 @@ archivePrefix = {arXiv},
   number={3-4},
   pages={171--186},
   year={2011},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.ocemod.2011.02.013}
 }
 
 @ARTICLE{Brueshaber:2019,
@@ -224,14 +228,15 @@ archivePrefix = {arXiv},
 }
 
 @article{Langton:2008b,
-  title={Hydrodynamic simulations of unevenly irradiated Jovian planets},
+  title={Hydrodynamic simulations of unevenly irradiated {Jovian} planets},
   author={Langton, Jonathan and Laughlin, Gregory},
   journal={The Astrophysical Journal},
   volume={674},
   number={2},
   pages={1106},
   year={2008},
-  publisher={IOP Publishing}
+  publisher={IOP Publishing},
+  doi={10.1086/523957}
 }
 @ARTICLE{Perez-Becker:2013,
        author = {{Perez-Becker}, Daniel and {Showman}, Adam P.},
@@ -260,7 +265,8 @@ archivePrefix = {arXiv},
   number={7825},
   pages={357--362},
   year={2020},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/s41586-020-2649-2}
 }
 
 @ARTICLE{Hunter:2007,
@@ -274,14 +280,15 @@ archivePrefix = {arXiv},
   doi={10.1109/MCSE.2007.55}}
 
   @article{Virtanen:2020,
-  title={SciPy 1.0: fundamental algorithms for scientific computing in Python},
+  title={SciPy 1.0: fundamental algorithms for scientific computing in {Python}},
   author={Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E and Haberland, Matt and Reddy, Tyler and Cournapeau, David and Burovski, Evgeni and Peterson, Pearu and Weckesser, Warren and Bright, Jonathan and others},
   journal={Nature methods},
   volume={17},
   number={3},
   pages={261--272},
   year={2020},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/s41592-019-0686-2}
 }
 
 @article{Gelb:2001,
@@ -291,7 +298,8 @@ archivePrefix = {arXiv},
   volume={129},
   number={9},
[…]
