### 2022-12-05 0e1d38d
Fix typo in paper.md
