+++ paper/paper.md
@@ -24,7 +24,7 @@ bibliography: paper.bib
 
 # Summary
 
-In order to answer questions about potentially habitability of exoplanets, it is important to develop a robust understanding of a variety of dynamic
+In order to answer questions about potential habitability of exoplanets, it is important to develop a robust understanding of a variety of dynamic
 processes that can take place in exoplanetary atmospheres. While many exoplanets are readily characterized with current facilities like Hubble and James Webb Space Telescope, exoplanet scientists work with indirect and limited observations of the planets that they study. To form hypotheses about their climate, weather, and atmospheric composition, astronomers need robust models that demonstrate how atmospheres act under different conditions. One-dimensional energy-balance models can capture complex mechanisms such as cloud formation and can rapidly explore the parameter ranges, but they fail to account for variations with longitude. In contrast, three-dimensional models capture the variation in latitude, longitude, and altitude, but they are computationally expensive, sometimes taking months to explore the parameter regimes. Their complexity can also obscure the mechanisms that govern atmospheric phenomena. This leaves a natural gap for two-dimensional models, which can capture the spatial variability as well as rapidly explore the parameter space and study the dynamical mechanisms.
 
 `SWAMPE` is a Python package for modeling the dynamics of exoplanetary atmospheres. `SWAMPE` is an intermediate-complexity, two-dimensional shallow-water general circulation model. Benchmarked for synchronously rotating hot Jupiters and sub-Neptunes, 
