+++ docs/source/conf.py
@@ -21,7 +21,7 @@ copyright = '2022, Ekaterina Landgren'
 author = 'Ekaterina Landgren'
 
 # The full version, including alpha/beta/rc tags
-release = '0.0.3'
+release = '1.0.0'
 
 
 # -- General configuration ---------------------------------------------------
