### 2022-12-08 #4872 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
